* Speed up model tests 20s -> 6.3s
* Speed up size field tests 8.5s -> 6.5s
* Speed up cache tests 30s -> 5s

Saves 41s total!